+++ index.html
@@ -230,6 +230,6 @@
     </div>
   </template>
   <!-- Подключение скрипта-->
-  <script src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
